# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  username: string;
  displayName: string;
}

export interface AuthResponse {
  user: AuthUser;
}

export interface Attendee {
  id: number;
  qrId: string;
  name: string;
  /** @nullable */
  email?: string | null;
  ticketType: string;
  /** @nullable */
  checkedInAt: string | null;
  createdAt: string;
}

export interface AttendeeInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  /** @minLength 1 */
  ticketType: string;
  qrId?: string;
}

export interface AttendeeImportInput {
  /** @minItems 1 */
  attendees: AttendeeInput[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  attendees: Attendee[];
}

export interface CheckInInput {
  /** @minLength 1 */
  qrId: string;
}

export type CheckInResultStatus = typeof CheckInResultStatus[keyof typeof CheckInResultStatus];


export const CheckInResultStatus = {
  valid: 'valid',
  duplicate: 'duplicate',
  invalid: 'invalid',
} as const;

export interface CheckInResult {
  status: CheckInResultStatus;
  message: string;
  attendee: Attendee | null;
}

export interface RecentCheckIn {
  name: string;
  ticketType: string;
  checkedInAt: string;
}

export interface DashboardSummary {
  total: number;
  checkedIn: number;
  remaining: number;
  recentCheckIns: RecentCheckIn[];
}

export type ListAttendeesParams = {
q?: string;
status?: ListAttendeesStatus;
};

export type ListAttendeesStatus = typeof ListAttendeesStatus[keyof typeof ListAttendeesStatus];


export const ListAttendeesStatus = {
  all: 'all',
  'checked-in': 'checked-in',
  pending: 'pending',
} as const;